+++ src/Graphics/GLU/Types.hs
@@ -1,0 +1,21 @@
+--------------------------------------------------------------------------------
+-- |
+-- Module      :  Graphics.GLU.Types
+-- Copyright   :  (c) Sven Panne 2016
+-- License     :  BSD3
+--
+-- Maintainer  :  Sven Panne <svenpanne@gmail.com>
+-- Stability   :  stable
+-- Portability :  portable
+--
+-- All (abstract) types from GLU 1.3, see <http://www.opengl.org/>.
+--
+--------------------------------------------------------------------------------
+
+module Graphics.GLU.Types (
+   GLUnurbs,
+   GLUquadric,
+   GLUtesselator
+) where
+
+import Graphics.GLU.TypesInternal